Rename `ebd-select` component to `ebd-input`
